pi_2023_3semestre/requirements.txt
anyio==3.6.2
asgiref==3.6.0
certifi==2023.5.7
click==8.1.3
Django==4.1.9
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.12
pydantic==1.10.7
PyJWT==2.7.0
pymongo==3.12.3
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
sniffio==1.3.0
sqlparse==0.2.4
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
